requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf-yasg
celery
redis
gunicorn
whitenoise
python-dotenv
pytest
pytest-django
coverage
flake8
django-cors-headers
